# Support plus sign before numbers