--- a/big.mjs
+++ b/big.mjs
@@ -68,7 +68,7 @@ var DP = 20,          // 0 to MAX_DP
   // The shared prototype object.
   P = {},
   UNDEFINED = void 0,
-  NUMERIC = /^-?(\d+(\.\d*)?|\.\d+)(e[+-]?\d+)?$/i;
+  NUMERIC = /^[+-]?(\d+(\.\d*)?|\.\d+)(e[+-]?\d+)?$/i;
 
 
 /*
@@ -141,6 +141,9 @@ function parse(x, n) {
 
   // Determine sign.
   x.s = n.charAt(0) == '-' ? (n = n.slice(1), -1) : 1;
+
+  // Strip sign.
+  if (n.charAt(0) == "+") n = n.slice(1);
 
   // Decimal point?
   if ((e = n.indexOf('.')) > -1) n = n.replace('.', '');