This adds support for parsing positive signed numbers like `+1`.