@Zebreus 2022-08-21
@MikeMcl Have you had a chance to look at the changes yet?

I use big.js to parse numbers from CSS stylesheets, but the CSS spec explicitly allows a positive sign. If that would be supported I can use it to directly parse the numbers from my stylesheets.