@johnmanko 2022-08-21
Although I don't have a direct need for this capability, it makes sense.  The `-` prefix is allowed, and so should `+`.