@MikeMcl 2022-08-21
See #59, #70 and [#134](https://github.com/MikeMcl/big.js/issues/134#issuecomment-589437410).

It's a reasonable suggestion, but I find it a little strange that you have gone straight to a PR when support for a leading plus sign is not likely to have been something I hadn't thought of previously but rather something that was deliberately not included. 

I am reluctant to accept this because it's another `if` statement in a very hot path and it's very easy for users who want to strip the plus sign to do it themselves anyway. What next? Support for leading whitespace? What about e.g. `"+ 1"?

One possibility is a `Big.from` or `Big.parse` method that is more liberal than the Big constructor. That could also allow underscores as in e.g. `"1_000_000"`.