@Zebreus 2022-08-23
Sorry, I hadn't looked through the closed issues. Should have done that before opening the PR. I assumed that a leading plus sign just wasn't a feature that many people required and nobody was missing it.

I totally agree to keep the number parsing scope of big.js to a minimum, after all it's a library for calculating numbers and not for parsing numbers. I think a well defined scope would be to support any decimal number string in a format that is also supported by the Number constructor, but nothing more. That is probably also what most users expect. It would also define that NumericLiteralSeparator and spaces are invalid, so there is no need to add support for those.
  
The performance impacts of adding support for a leading plus sign are basically zero. I did some benchmarking and could not find any significant performance difference between supporting plus or not.

<details>
  <summary>Detailed benchmark results</summary>

I tested the performance of the Big function with the following test function:

```javascript
const test = (value) => {
  const a = new Date()
  for(var i = 0; i < 1e8; ++i) { Big(value) }
  const b = new Date()
  return b.getTime() - a.getTime()
}
```

I tested my fork against the current master with three different values

| value | time without plus support | time with plus support|
|-|-|-|
| `8` | 8325ms | 8186ms
| `-998` | 10663ms | 10585ms
| `845439825732e67` | 19194ms | 19254ms

</details>

I think there are enough users who want a Big constructor that is more analogous to the Number constructor, to warrant a single line of code more.