@MikeMcl 2022-08-23
> I assumed that a leading plus sign just wasn't a feature that many people required and nobody was missing it.

I agree with your assumption that it isn't a feature that many people require.

> I think there are enough users who want a Big constructor that is more analogous to the Number constructor, to warrant a single line of code more.

Hmmm... this seems to contradict your above assumption somewhat, and your PR involves more than adding a single line. 

> I think a well defined scope would be to support any decimal number string in a format that is also supported by the Number constructor, but nothing more... It would also define that... spaces are invalid,

Well, we can have, for example

```js
Number('  \n   +2  \t\n  ')    // 2
```

so I suppose you mean whatever is supported by `Number` apart from leading and trailing whitespace characters.

Other arguments *for* are that a leading plus sign is supported by sister libraries *decimal.js* and *bignumber.js*, and also by Java's BigDecimal and Python's Decimal, and this library already supports the similarly redundant plus sign for exponents.

```js
Big('1e+2').toString();    // '100'
```

It's a borderline decision (that has already been made and remade previously) and in such cases here I usually follow the maxim: 'If in doubt, leave it out'.

Anyway, thanks for your efforts. I may include this but I'm not deciding the matter today.