@Zebreus 2022-09-13
Closing this PR, because it probably does not fit into big.js. I have published this as [`@zebreus/big.js`](https://www.npmjs.com/package/@zebreus/big.js) to npm, if anyone else needs support for plus signs.

Thanks for the creating this library, I really enjoy using it.